requirements.txt
absl-py==1.3.0
altair==4.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
debugpy==1.6.3
decorator==5.1.1
deepface==0.0.75
dnspython==2.2.1
easy-colors==0.0.1
entrypoints==0.4
executing==1.0.0
extra-streamlit-components==0.1.56
filelock==3.8.2
fire==0.4.0
Flask==2.2.2
flatbuffers==22.12.6
gast==0.4.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.15.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mtcnn==0.1.1
nest-asyncio==1.5.5
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.1
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rich==12.6.0
rsa==4.9
semver==2.13.0
send-mail==0.0.2
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.5.0
streamlit==1.15.2
streamlit-authenticator==0.1.5
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.11.0
